Lemon Car Prediction Model
The motive of this study was to analyze whether the car dealerships
could nip the problem of buying ‘lemon’ in an auction in the bud by
leveraging the power of data analytics to make informed decisions at
the time of car purchase.

● Modeled in R, Tableau and MS Azure to provide a real business value
to a car dealership dedicated towards providing the best inventory
selection possible to the customers for all car brands across 50 states
containing 19 predictors.

● Determined Decision Jungle as the best model and further optimized
the model with prediction accuracy of 90.4%

● Performed analysis in four sequential phases – Monetary Impact, Data
Exploration, Data Modeling and Prediction analysis across 10k records
reducing processing time by 20%.